feat: alter exercises table Mentions: #85
